api/requirements.txt
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
django-cros-headers==3.14.0